requirements.txt
aiocontextvars==0.2.2
aiofiles==0.7.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
contextvars==2.4
greenlet==1.1.0
h11==0.9.0
httpcore==0.11.1
httptools==0.2.0
httpx==0.15.4
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.0.1
loguru==0.5.3
multidict==5.1.0
requests==2.25.1
rfc3986==1.5.0
sanic==20.12.3
sniffio==1.2.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
websockets==8.1
win32-setctime==1.0.3
yarl==1.6.3
zipp==3.4.1
pyjwt==2.0.0
sanic-jwt==1.6.0
pillow==8.2.0